requirements/current.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.16.0
psycopg2-binary==2.9.3
PyJWT==2.8.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.3
twilio==7.15.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==2.0.7


